Refactor
Move init Haptizer and Speaker to CameraFragment from MainActivity. Thanks to that we don't need to refer to MainActivity variables.
