+++ Android/PolishBanknotesApp/app/src/client/java/pg/eti/project/polishbanknotes/accessability/Beeper.kt
@@ -1,4 +1,4 @@
-package pg.eti.project.polishbanknotes.accesability
+package pg.eti.project.polishbanknotes.accessability
 
 import android.content.Context
 import android.content.Context.AUDIO_SERVICE
